+++ src/components/NotificationRequest.svelte
@@ -2,32 +2,33 @@
   import { browser } from '$app/env'
   import Notification16 from 'carbon-icons-svelte/lib/Notification16'
 
-  let notificationsAreAllowed = false
-  let browserSupportsNotifications = false
+  let notificationsAreAllowed = NaN
+  let browserSupportsNotifications = NaN
 
   function requestPermission () {
     Notification.requestPermission().then(function (status) {
-      if (status === 'granted') {
-        notificationsAreAllowed = true
-      }
+      notificationsAreAllowed = (status === 'granted')
     })
   }
 
   if (browser) {
     browserSupportsNotifications = ('Notification' in window)
-
-    if (Notification.permission === 'granted') {
-      notificationsAreAllowed = true
-    }
+    notificationsAreAllowed = (Notification.permission === 'granted')
   }
+
+  let loaded = false
+
+  $: loaded = !(isNaN(notificationsAreAllowed) && isNaN(browserSupportsNotifications))
 </script>
 
 <section>
-  {#if !browserSupportsNotifications}
+  {#if !loaded }
+        // not ready
+  {:else if !browserSupportsNotifications}
     <div class="notification mt-2 is-warning is-light">
       <p class="">This browser does not support notifications.</p>
     </div>
-  {:else if !notificationsAreAllowed}
+  {:else if !notificationsAreAllowed }
     <div class="notification mt-2 is-warning is-light">
       <button class="button is-success is-outlined is-small is-pulled-right" on:click={requestPermission}>
         <span class="icon is-small">