+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.0.2] - 2022-08-12
+### Changed
+- break and work are in different style
+- improve ui for notification support check
+- other small ui improvements
+
 ## [0.0.1] - 2022-01-22
 ### Added
 - basic funcionality of Mini and Regular breaks
@@ -12,5 +18,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - simple Preferences (intervals and durations for Mini and Regular breaks)
 - debug information
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.0.1...HEAD
-[0.0.1]: https://github.com/olivierlacan/keep-a-changelog/releases/tag/v0.0.1
+[Unreleased]: https://github.com/hovancik/stretchly-for-web/compare/v0.0.1...HEAD
+[0.0.1]: https://github.com/hovancik/stretchly-for-web/releases/tag/v0.0.1